URI::parse unescaped source to re-encode
Instead of checking if it is a shell safe url, unencode it as it was
previously being tested, and then allow parse to attempt to re-encode it
for our use using it's to_s method.  This seems to pass all tests minus
two that are not expecting single quotes to be encoded, but in my tests,
urls placed in browser and in PDFKit.new.to_pdf seem to yield the same
results.

I did not want to change existing tests until getting feedback from
others before doing so as I do not want to cause unexpected results from
use, which this seems to do.

+++ spec/source_spec.rb
@@ -102,32 +102,56 @@ describe PDFKit::Source do
 
     it "should not allow backtick shell execution in url" do
       filename = Dir::Tmpname.create('backtick_file') { |path| path }
+      File.delete(filename) if File.file?(filename)
 
-      if File.file?(filename)
-        File.delete(filename)
-      end
       source = PDFKit::Source.new("http://example.com/?name={'%20`sleep 5`'}")
-      expect(source.to_input_for_command).to eq "\"http://example.com/\\?name\\=\\{\\'\\%20\\`sleep\\ 5\\`\\'\\}\""
+      expect(source.to_input_for_command).to eq "\"http://example.com/?name={%27%20%60sleep%205%60%27}\""
 
       begin
         PDFKit.new("http%20`touch #{filename}`").to_pdf
-      rescue PDFKit::ImproperWkhtmltopdfExitStatus
+      rescue URI::InvalidURIError
       end
       expect(File.file?(filename)).to eq false
     end
 
     it "should not allow $( shell execution in url" do
       filename = Dir::Tmpname.create('dolar_sign_file') { |path| path }
+      File.delete(filename) if File.file?(filename)
 
-      if File.file?(filename)
-        File.delete(filename)
-      end
       source = PDFKit::Source.new("http://example.com/?name={'%20$(sleep 5)'}")
-      expect(source.to_input_for_command).to eq "\"http://example.com/\\?name\\=\\{\\'\\%20\\$\\(sleep\\ 5\\)\\'\\}\""
+      expect(source.to_input_for_command).to eq "\"http://example.com/?name={%27%20$(sleep%205)%27}\""
 
       begin
-      PDFKit.new("http%20$(touch #{filename})").to_pdf
-      rescue PDFKit::ImproperWkhtmltopdfExitStatus
+        PDFKit.new("http%20$(touch #{filename})").to_pdf
+      rescue URI::InvalidURIError
+      end
+      expect(File.file?(filename)).to eq false
+    end
+
+    it "should not allow || shell execution in url" do
+      filename = Dir::Tmpname.create('or_file') { |path| path }
+      File.delete(filename) if File.file?(filename)
+
+      source = PDFKit::Source.new("http://%20a\" || sleep 3; \"")
+      expect { source.to_input_for_command }.to raise_exception(URI::InvalidURIError)
+
+      begin
+        PDFKit.new("http://%20a\" || touch #{filename}); \"").to_pdf
+      rescue URI::InvalidURIError
+      end
+      expect(File.file?(filename)).to eq false
+    end
+
+    it "should not allow && shell execution in url" do
+      filename = Dir::Tmpname.create('and_file') { |path| path }
+      File.delete(filename) if File.file?(filename)
+
+      source = PDFKit::Source.new("http://%20a\" && sleep 3; \"")
+      expect { source.to_input_for_command }.to raise_exception(URI::InvalidURIError)
+
+      begin
+        PDFKit.new("http://%20a\" && touch #{filename}); \"").to_pdf
+      rescue URI::InvalidURIError
       end
       expect(File.file?(filename)).to eq false
     end
